Change sbrk() to take an int rather than an intptr_t (following the OpenBSD way rather than the Linux way; various non-C bits of the ACK assume it takes an int, so it's cleaner).

+++ plat/linux/libsys/sbrk.c
@@ -22,7 +22,7 @@ int brk(void* end)
 	return e;
 }
 
-void* sbrk(intptr_t increment)
+void* sbrk(int increment)
 {
 	char* old;
 	char* new;
